extracted reference processing to a new class
